Factor out cli-options module
